composite iterator
