+++ airbyte-commons/src/main/java/io/airbyte/commons/util/AutoCloseableIterators.java
@@ -26,21 +26,13 @@ package io.airbyte.commons.util;
 
 import com.google.common.collect.Iterators;
 import io.airbyte.commons.concurrency.VoidCallable;
-import io.airbyte.commons.concurrency.VoidCallableNoException;
-import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.function.Function;
 import java.util.function.Supplier;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 public class AutoCloseableIterators {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(AutoCloseableIterators.class);
-
   /**
    * Coerces a vanilla {@link Iterator} into a {@link AutoCloseableIterator} by adding a no op close
    * function.
@@ -62,7 +54,7 @@ public class AutoCloseableIterators {
    * @param <T> type
    * @return new autocloseable iterator with the close function appended
    */
-  public static <T> AutoCloseableIterator<T> fromIterator(Iterator<T> iterator, VoidCallableNoException onClose) {
+  public static <T> AutoCloseableIterator<T> fromIterator(Iterator<T> iterator, VoidCallable onClose) {
     return new DefaultAutoCloseableIterator<>(iterator, onClose::call);
   }
 
@@ -93,34 +85,6 @@ public class AutoCloseableIterators {
   }
 
   /**
-   * Returns a {@link AutoCloseableIterator} that will call {@link AutoCloseableIterator#close()} at most one
-   * time. Either as soon as {@link Iterator#hasNext()} returns false for the first time or when
-   * {@link AutoCloseableIterator#close()} is called,
-   *
-   * @param autoCloseableIterator wrapped autocloseable iterator
-   * @param <T> type
-   * @return autocloseable iterator
-   */
-  public static <T> AutoCloseableIterator<T> decorateWithEagerClose(AutoCloseableIterator<T> autoCloseableIterator) {
-    return new EagerCloseIterator<>(autoCloseableIterator);
-  }
-
-  /**
-   * Returns a {@link AutoCloseableIterator} that is composed of a {@link DefaultAutoCloseableIterator},
-   * {@link LazyAutoCloseableIterator}, and {@link EagerCloseIterator}.
-   *
-   * @param streamSupplier supplies the stream this supplier will be called one time.
-   * @param <T> type
-   * @return autocloseable iterator
-   */
-  public static <T> AutoCloseableIterator<T> lazyEagerCloseAutoCloseableIterator(Supplier<Stream<T>> streamSupplier) {
-    return decorateWithEagerClose(lazyIterator(() -> {
-      final Stream<T> stream = streamSupplier.get();
-      return fromStream(stream);
-    }));
-  }
-
-  /**
    * Append a function to be called on {@link AutoCloseableIterator#close}.
    *
    * @param autoCloseableIterator autocloseable iterator to add another close to
@@ -128,7 +92,7 @@ public class AutoCloseableIterators {
    * @param <T> type
    * @return new autocloseable iterator with the close function appended
    */
-  public static <T> AutoCloseableIterator<T> appendOnClose(AutoCloseableIterator<T> autoCloseableIterator, VoidCallableNoException voidCallable) {
+  public static <T> AutoCloseableIterator<T> appendOnClose(AutoCloseableIterator<T> autoCloseableIterator, VoidCallable voidCallable) {
     return new DefaultAutoCloseableIterator<>(autoCloseableIterator, () -> {
       autoCloseableIterator.close();
       voidCallable.call();
@@ -164,55 +128,8 @@ public class AutoCloseableIterators {
     return new DefaultAutoCloseableIterator<>(iteratorCreator.apply(autoCloseableIterator), autoCloseableIterator::close);
   }
 
-  /**
-   * Concatenates {@link AutoCloseableIterator}.
-   *
-   * @param iterators iterators to concatenate.
-   * @param <T> type
-   * @return concatenated iterator
-   */
-  public static <T> AutoCloseableIterator<T> concat(List<AutoCloseableIterator<T>> iterators) {
-    return new DefaultAutoCloseableIterator<>(Iterators.concat(iterators.iterator()), () -> iteratorCloseFunction(iterators));
-  }
-
   public static <T> CompositeIterator<T> concatWithEagerClose(List<AutoCloseableIterator<T>> iterators) {
-    return DefaultCompositeIterator.fromAutoCloseableIterators(iterators);
-  }
-
-  public static class DefaultCompositeIterator<T> extends DefaultAutoCloseableIterator<T> implements CompositeIterator<T> {
-
-    private DefaultCompositeIterator(AutoCloseableIterator<T> iterator, VoidCallable onClose) {
-      super(iterator, onClose);
-    }
-
-    public static <T> CompositeIterator<T> fromAutoCloseableIterators(List<AutoCloseableIterator<T>> iterators) {
-      return new DefaultCompositeIterator<>(concatWithEagerClose(iterators), () -> iteratorCloseFunction(iterators));
-    }
-
-    public static <T> AutoCloseableIterator<T> concatWithEagerClose(List<AutoCloseableIterator<T>> iterators) {
-      final List<AutoCloseableIterator<T>> eagerCloseIterators = iterators
-          .stream()
-          .map(EagerCloseIterator::new)
-          .collect(Collectors.toList());
-
-      return AutoCloseableIterators.concat(eagerCloseIterators);
-    }
-  }
-
-  private static <T> void iteratorCloseFunction(List<AutoCloseableIterator<T>> iterators) throws Exception {
-    final List<Exception> exceptions = new ArrayList<>();
-    for (AutoCloseableIterator<T> iterator : iterators) {
-      try {
-        iterator.close();
-      } catch (Exception e) {
-        LOGGER.error("exception while closing", e);
-        exceptions.add(e);
-      }
-    }
-
-    if (!exceptions.isEmpty()) {
-      throw exceptions.get(0);
-    }
+    return new CompositeIterator<>(iterators);
   }
 
 }
